Move loadPrivateKey into cpp
